backend/requirements.txt
fastapi==0.115.6
sqlalchemy==2.0.37
uvicorn==0.34.0
pydantic==2.10.5
python-dotenv==1.0.1
pymysql==1.1.1
email-validator==2.2.0
pytz
pydantic-settings==2.1.0